% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fashion_train}
\alias{fashion_train}
\title{Reproduced from fashion-mnist data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1000 rows and 785 columns.}
\source{
\url{https://github.com/ymattu/fashion-mnist-csv}
}
\usage{
fashion_train
}
\description{
Reproduced from fashion-mnist data
}
\keyword{datasets}
