% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris_test}
\alias{iris_test}
\title{even-numbered rows of iris data}
\format{An object of class \code{data.frame} with 75 rows and 5 columns.}
\usage{
iris_test
}
\description{
even-numbered rows of iris data
}
\keyword{datasets}
