% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_cv_opt.R
\name{svm_cv_opt}
\alias{svm_cv_opt}
\title{Bayesian Optimization for SVM (Cross Validiation)}
\usage{
svm_cv_opt(data, label, gamma_range = c(10^(-3), 10^1), c_range = c(10^(-2),
  10^2), svm_kernel = "radial", degree = 3, n_folds = 0,
  init_points = 4, n_iter = 10, acq = "ei", kappa = 2.576, eps = 0,
  optkernel = list(type = "exponential", power = 2))
}
\arguments{
\item{data}{data}

\item{label}{label for classification}

\item{gamma_range}{The range of gamma. Default is c(10 ^ (-3), 10 ^ 1)}

\item{c_range}{The range of C(Cost). Deafult is c(10 ^ (-2), 10 ^ 2)}

\item{svm_kernel}{Kernel used in SVM. You might consider changing some of the following parameters, depending on the kernel type.
\itemize{
  \item \strong{linear:} \eqn{u'v}
  \item \strong{polynomial:} \eqn{(\gamma u'v +coef0)^{degree}}
  \item \strong{radial basis:} \eqn{\exp(-\gamma|u-v|^2)}
  \item \strong{sigmoid:} \eqn{tanh(\gamma u'v + coef0)}
}}

\item{degree}{Parameter needed for kernel of type polynomial (default: 3)}

\item{n_folds}{if a integer value k>0 is specified, a k-fold cross validation on the training data is performed to assess the quality of the model: the accuracy rate for classification and the Mean Squared Error for regression}

\item{init_points}{Number of randomly chosen points to sample the
target function before Bayesian Optimization fitting the Gaussian Process.}

\item{n_iter}{Total number of times the Bayesian Optimization is to repeated.}

\item{acq}{Acquisition function type to be used. Can be "ucb", "ei" or "poi".
\itemize{
  \item \code{ucb} GP Upper Confidence Bound
  \item \code{ei} Expected Improvement
  \item \code{poi} Probability of Improvement
}}

\item{kappa}{tunable parameter kappa of GP Upper Confidence Bound, to balance exploitation against exploration,
increasing kappa will make the optimized hyperparameters pursuing exploration.}

\item{eps}{tunable parameter epsilon of Expected Improvement and Probability of Improvement, to balance exploitation against exploration,
increasing epsilon will make the optimized hyperparameters are more spread out across the whole range.}

\item{optkernel}{Kernel (aka correlation function) for the underlying Gaussian Process. This parameter should be a list
that specifies the type of correlation function along with the smoothness parameter. Popular choices are square exponential (default) or matern 5/2}
}
\value{
The test accuracy and a list of Bayesian Optimization result is returned:
\itemize{
  \item \code{Best_Par} a named vector of the best hyperparameter set found
  \item \code{Best_Value} the value of metrics achieved by the best hyperparameter set
  \item \code{History} a \code{data.table} of the bayesian optimization history
  \item \code{Pred} a \code{data.table} with validation/cross-validation prediction for each round of bayesian optimization history
}
}
\description{
This function estimates parameters for SVM(Gaussian Kernel) based on bayesian optimization
}
\examples{
\dontrun{
library(MlBayesOpt)
# This takes a lot of time
res0 <- svm_cv_opt(data = fashion,
                   label = y,
                   n_folds = 10)
}

}
