% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_mnist}
\alias{test_mnist}
\title{produced from kaggle competition}
\format{An object of class \code{data.frame} with 1000 rows and 785 columns.}
\source{
\url{https://github.com/ozt-ca/tjo.hatenablog.samples/tree/master/r_samples/public_lib/jp/mnist_reproduced}
}
\usage{
test_mnist
}
\description{
produced from kaggle competition
}
\keyword{datasets}
