% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_opt.R
\name{svm_opt}
\alias{svm_opt}
\title{Bayesian Optimization for SVM (Gaussian Kernel)}
\usage{
svm_opt(train_data, train_label, test_data, test_label,
  gamma_range = c(10^(-5), 10^5), c_range = c(10^(-2), 10^2),
  init_points = 20, n_iter = 1, acq = "ei", kappa = 2.576, eps = 0,
  kernel = list(type = "exponential", power = 2))
}
\arguments{
\item{train_data}{A data frame for training of xgboost}

\item{train_label}{The column of class to classify in the training data}

\item{test_data}{A data frame for training of xgboos}

\item{test_label}{The column of class to classify in the test data}

\item{gamma_range}{The range of gamma. Default is c(10 ^ (-5), 10 ^ 5)}

\item{c_range}{The range of C(Cost). Deafult is c(10 ^ (-2), 10 ^ 2)}

\item{init_points}{Number of randomly chosen points to sample the
target function before Bayesian Optimization fitting the Gaussian Process.}

\item{n_iter}{Total number of times the Bayesian Optimization is to repeated.}

\item{acq}{Acquisition function type to be used. Can be "ucb", "ei" or "poi".
\itemize{
  \item \code{ucb} GP Upper Confidence Bound
  \item \code{ei} Expected Improvement
  \item \code{poi} Probability of Improvement
}}

\item{kappa}{tunable parameter kappa of GP Upper Confidence Bound, to balance exploitation against exploration,
increasing kappa will make the optimized hyperparameters pursuing exploration.}

\item{eps}{tunable parameter epsilon of Expected Improvement and Probability of Improvement, to balance exploitation against exploration,
increasing epsilon will make the optimized hyperparameters are more spread out across the whole range.}

\item{kernel}{Kernel (aka correlation function) for the underlying Gaussian Process. This parameter should be a list
that specifies the type of correlation function along with the smoothness parameter. Popular choices are square exponential (default) or matern 5/2}
}
\value{
The test accuracy and a list of Bayesian Optimization result is returned:
\itemize{
  \item \code{Best_Par} a named vector of the best hyperparameter set found
  \item \code{Best_Value} the value of metrics achieved by the best hyperparameter set
  \item \code{History} a \code{data.table} of the bayesian optimization history
  \item \code{Pred} a \code{data.table} with validation/cross-validation prediction for each round of bayesian optimization history
}
}
\description{
This function estimates parameters for SVM(Gaussian Kernel) based on bayesian optimization
}
